refactor: Move components to ui-kit directory
- Create ui-kit directory for common components
- Move Button and Dropdown components to ui-kit
- Update import paths in index.tsx
- Update Tailwind config to include ui-kit path
