+++ tailwind.config.js
@@ -3,6 +3,7 @@ module.exports = {
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx}",
+    "./src/ui-kit/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {
